src/backend/requirements.txt
# PersBot Backend Dependencies

# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Audio
faster-whisper>=0.10.0
pvporcupine>=4.0.0
pyttsx3>=2.90
numpy>=1.24.0
sounddevice>=0.4.6

# AI
ollama>=0.1.0
openai>=1.3.0

# Windows Automation
pywinauto>=0.6.8
pyautogui>=0.9.54

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.2.0

# MCP - Model Context Protocol
mcp>=1.0.0

# Memory
rank-bm25>=0.2.2
numpy>=1.24.0

# Channels
requests>=2.31.0
httpx>=0.25.0
lark-oapi>=1.0.0
